added wordlist and fetch only active hosts
